requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-classic
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
tiktoken
